splitting classifier logic
