+++ folktexts/classifier/__init__.py
@@ -1,0 +1,3 @@
+from .base import LLMClassifier                                 # noqa: F401
+from .transformers_classifier import TransformersLLMClassifier  # noqa: F401
+from .web_api_classifier import WebAPILLMClassifier             # noqa: F401
